+++ src/components/field/cell.tsx
@@ -1,9 +1,11 @@
-import React from 'react';
+import React, { useState } from 'react';
 import './field.scss';
 import { useSelector } from 'react-redux';
+import { Sound } from '@components';
 import { IAppState } from '@types';
 import { DEFAULT_FIELD_SIDE_SIZE, MOBILE_FIELD_SIDE_SIZE, MOBILE_WINDOW_SIZE, Theme } from '@constants';
 import { useWindowSize } from '@hooks';
+import musicfile from '../../assets/sound/click.mp3';
 
 type CellProps = {
   isActive: boolean;
@@ -14,10 +16,15 @@ type CellProps = {
 };
 
 export const Cell = ({ isActive, isSelected, isFocused, letter, handleSelectedCell }: CellProps): JSX.Element => {
+  const [isPlay, setIsPlay] = useState<boolean>(false);
+
   const size = useWindowSize();
   const fieldSize = useSelector((state: IAppState) => state.game.fieldSize);
   const themeInitial = useSelector((state: IAppState) => state.settings.currentTheme);
+  const isSoundMuteOn = useSelector((state: IAppState) => state.settings.isSoundOn);
   const themeChange = themeInitial === Theme.light ? 'cell-light' : 'cell-dark';
+
+  const handlePlay = () => setIsPlay(true);
 
   const getClassName = () => {
     const classes = [`cell ${themeChange}`];
@@ -41,7 +48,17 @@ export const Cell = ({ isActive, isSelected, isFocused, letter, handleSelectedCe
   };
 
   return (
-    <div className={getClassName()} style={style} onClick={handleSelectedCell} role="button" tabIndex={-1}>
+    <div
+      className={getClassName()}
+      style={style}
+      onClick={() => {
+        handleSelectedCell();
+        handlePlay();
+      }}
+      role="button"
+      tabIndex={-1}
+    >
+      <Sound src={musicfile} playing={isPlay} format={['mp3']} loop={false} mute={!isSoundMuteOn} onEnd={setIsPlay} />
       {letter}
     </div>
   );