+++ src/custom.d.ts
@@ -12,3 +12,8 @@ declare module '*.jpg' {
   const content: string;
   export default content;
 }
+
+declare module '*.mp3' {
+  const content: string;
+  export default content;
+}